+++ uk/co/petertribble/sphaero2/QuadCutter.java
@@ -80,7 +80,7 @@ public class QuadCutter extends JigsawCutter {
 	 */
 	int rows = (int) Math.round(
 		Math.sqrt(widthToHeightRatio * prefPieces * height / width));
-	int columns = (int) Math.round(prefPieces / rows);
+	int columns = Math.round(prefPieces / rows);
 
 	jp.setMaximum(rows*columns);
 	jp.setValue(iprogress);
